requirements.txt
chainlit==1.2.0
firecrawl_py==1.2.4
google_search_results==2.4.2
langchain==0.3.1
langchain_community==0.3.1
langchain_core==0.3.6
langchain_huggingface==0.1.0
langchain_openai==0.2.1
langsmith==0.1.129
llama_index==0.11.14
openai==1.50.2
python-dotenv==1.0.1
serpapi==0.1.5
tiktoken==0.7.0
